+++ lib/ui/profile/ProfileStoredDataPanel.dart
@@ -3,24 +3,18 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:http/http.dart';
-import 'package:illinois/model/CustomCourses.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Appointments.dart';
+import 'package:illinois/service/Assistant.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Canvas.dart';
-import 'package:illinois/service/CustomCourses.dart';
+import 'package:illinois/service/Gateway.dart';
 import 'package:illinois/service/Identity.dart';
-import 'package:illinois/service/IlliniCash.dart';
 import 'package:illinois/service/Occupations.dart';
-import 'package:illinois/service/RecentItems.dart';
 import 'package:illinois/service/Rewards.dart';
+import 'package:illinois/service/Transportation.dart';
 import 'package:illinois/service/Wellness.dart';
-import 'package:illinois/service/WellnessRings.dart';
-import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/LinkButton.dart';
-import 'package:rokwire_plugin/model/event2.dart';
-import 'package:rokwire_plugin/model/poll.dart';
 import 'package:rokwire_plugin/service/events2.dart';
 import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/inbox.dart';
@@ -32,7 +26,7 @@ import 'package:rokwire_plugin/service/surveys.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class ProfileStoredDataPanel extends StatefulWidget {
-  static const String notifyRefresh  = "edu.illinois.rokwire.home.refresh";
+  static const String _notifyRefresh  = "edu.illinois.rokwire.home.refresh";
 
   ProfileStoredDataPanel._();
 
@@ -56,78 +50,38 @@ class ProfileStoredDataPanel extends StatefulWidget {
 }
 
 
-typedef _StoredDataProvider = Future<String?> Function();
+typedef _UserDataProvider = Future<Map<String, dynamic>?> Function();
 
-enum _StoredDataType {
-  // rokwire.illinois.edu/core
-  coreAccount,
+enum _StoredDataSource {
+  core,
+  identity,
+  rewards,
+  notifications,
+  social,
 
-  // rokwire.illinois.edu/notifications
-  notificationsAccount,
+  calendar,
+  groups,
+  polls,
+  surveys,
 
-  // rokwire.illinois.edu/lms
-  canvasAccount,
-  customCourses,
-  customCoursesHistory,
+  lms,
+  appointments,
+  occupations,
 
-  // rokwire.illinois.edu/calendar
-  myEvents,
-  participatedEvents, // TBD: registered or attended events
-
-  // rokwire.illinois.edu/gr
-  myGroups,
-  myGroupsPosts, // TBD: all my posts and messages in all groups
-  myGroupsStats,
-
-  // rokwire.illinois.edu/polls
-  myPools,
-  participatedPolls,
-
-  // rokwire.illinois.edu/surveys
-  mySurveys,
-  participatedSurveys,
-
-  // rokwire.illinois.edu/identity
-  myStudentId,
-  myStudentClassification,
-  myMobileCredentials,
-
-  // rokwire.illinois.edu/rewards
-  myRewardsBalance,
-  myRewardsHistory,
-
-  // rokwire.illinois.edu/wellness
-  myWellnessToDoCategories,
-  myWellnessToDoItems,
-  myWellnessRingsRecords,
-
-  // rokwire.illinois.edu/appointments
-  myAppointments,
-
-  // rokwire.illinois.edu/occupations
-  myOccupationMatches,
+  transportation,
+  wellness,
+  assistant,
+  gateway,
 
   // icard.uillinois.edu
-  iCard,
-
   // housing.illinois.edu
-  studentSummary,
-  illiniCashBalance,
-  illiniCashTransactions,
-  mealPlanTransactions,
-  cafeCreditTransactions,
-
-  // recent_items
-  recentItems
+  // storage
 }
 
 class _ProfileStoredDataPanelState extends State<ProfileStoredDataPanel> {
 
   final StreamController<String> _updateController = StreamController.broadcast();
-  final Map<_StoredDataType, GlobalKey> _storedDataKeys = <_StoredDataType, GlobalKey>{};
-
-  List<UserCourse>? _userCourses;
-  Set<Completer<Response?>>? _userCoursesListeners;
+  final Map<_StoredDataSource, GlobalKey<_ProfileStoredDataWidgetState>> _storedDataKeys = <_StoredDataSource, GlobalKey<_ProfileStoredDataWidgetState>>{};
 
   @override
   void initState() {
@@ -161,368 +115,175 @@ class _ProfileStoredDataPanelState extends State<ProfileStoredDataPanel> {
 
   Widget get _panelContent => Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
     Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-      ..._coreContent,
-      ..._notificationsContent,
-      ..._lmsContent,
-      ..._calendarContent,
-      ..._groupsContent,
-      ..._pollsContent,
-      ..._surveysContent,
-      ..._identityContent,
-      ..._rewardsContent,
-      ..._wellnessContent,
-      ..._appointmentsContent,
-      ..._occupationsContent,
-      ..._icardContent,
-      ..._housingContent,
-      ...recentItemsContent,
+      _coreContent,
+      _identityContent,
+      _rewardsContent,
+      _notificationsContent,
+      _socialContent,
+
+      _calendarContent,
+      _groupsContent,
+      _pollsContent,
+      _surveysContent,
+
+      _lmsContent,
+      _appointmentsContent,
+      _occupationsContent,
+
+      _transportationContent,
+      _wellnessContent,
+      _assistantContent,
+      _gatewayContent,
+
+      //..._icardContent,
+      //..._housingContent,
+      //...recentItemsContent,
     ]),
   );
 
-  // rokwire.illinois.edu/core
-
-  List<Widget> get _coreContent => <Widget>[
+  Widget get _coreContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.coreAccount] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.core_account.title', "Core Account"),
-      dataProvider: _provideCoreAccountJson,
+      key: _storedDataKeys[_StoredDataSource.core] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.core,
+      dataProvider: () => Auth2().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.source.core.title', "Core"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideCoreAccountJson() async          => _provideResponseData(await Auth2().loadAccountResponse());
-
-  // rokwire.illinois.edu/notifications
-
-  List<Widget> get _notificationsContent => <Widget>[
+  Widget get _identityContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.notificationsAccount] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.notifications_user.title', "Notifications Account"),
-      dataProvider: _provideNotificationsAccountJson,
+      key: _storedDataKeys[_StoredDataSource.identity] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.identity,
+      dataProvider: () => Identity().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.source.identity.title', "Identity"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideNotificationsAccountJson() async => _provideResponseData(await Inbox().loadUserInfoResponse());
-
-  // rokwire.illinois.edu/lms
-
-  List<Widget> get _lmsContent => <Widget>[
+  Widget get _rewardsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.canvasAccount] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.canvas_user.title', "Canvas Account"),
-      dataProvider: _provideCanvasAccountJson,
+      key: _storedDataKeys[_StoredDataSource.rewards] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.rewards,
+      dataProvider: () => Rewards().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.source.rewards.title', "Rewards"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _notificationsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.customCourses] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.custom_courses.title', "My Canvas Courses"),
-      dataProvider: _provideCustomCoursesJson,
+      key: _storedDataKeys[_StoredDataSource.notifications] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.notifications,
+      dataProvider: () => Inbox().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.notifications.title', "Notifications"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _socialContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.customCoursesHistory] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.custom_courses_history.title', "My Canvas Courses Content"),
-      dataProvider: _provideCustomCoursesHistoryJson,
+      key: _storedDataKeys[_StoredDataSource.social] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.social,
+      dataProvider: () => Social().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.social.title', "Social"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideCanvasAccountJson() async        => _provideResponseData(await Canvas().loadSelfUserResponse());
-
-  //Future<String?> _provideCustomCoursesJson() async        => _provideResponseData(await CustomCourses().loadUserCoursesResponse());
-  Future<String?> _provideCustomCoursesJson() async {
-    Response? response;
-    if (_userCoursesListeners != null) {
-      Completer<Response?> completer = Completer<Response?>();
-      _userCoursesListeners?.add(completer);
-      response = await completer.future;
-    }
-    else {
-      Set<Completer<Response?>> completers = <Completer<Response?>>{};
-      _userCoursesListeners = completers;
-      response = await CustomCourses().loadUserCoursesResponse();
-      _userCourses = _buildUserCourses(response);
-      _userCoursesListeners = null;
-      for (Completer<Response?> completer in completers) {
-        completer.complete();
-      }
-    }
-    return _provideResponseData(response);
-  }
-
-  //Future<String?> _provideCustomCoursesHistoryJson() async => _provideResponseData(await CustomCourses().loadUserContentHistoryResponse());
-  Future<String?> _provideCustomCoursesHistoryJson() async {
-    if (_userCoursesListeners != null) {
-      Completer<Response?> completer = Completer<Response?>();
-      _userCoursesListeners?.add(completer);
-      await completer.future;
-    }
-    Response? response = await CustomCourses().loadUserContentHistoryResponse(ids: _userCoursesIds);
-    return _provideResponseData(response);
-  }
-
-  List<String> get _userCoursesIds => List<String>.from(_userCourses?.map((UserCourse course) => course.id) ?? []);
-
-  List<UserCourse>? _buildUserCourses(Response? response) {
-    List<dynamic>? userCoursesJson = (response?.statusCode == 200) ? JsonUtils.decodeList(response?.body) : null;
-    return (userCoursesJson != null) ? UserCourse.listFromJson(userCoursesJson) : null;
-  }
-
-  // rokwire.illinois.edu/calendar
-
-  List<Widget> get _calendarContent => <Widget>[
+  Widget get _calendarContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myEvents] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_events.title', "My Events"),
-      dataProvider: _provideMyEventsJson,
+      key: _storedDataKeys[_StoredDataSource.calendar] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.calendar,
+      dataProvider: () => Events2().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.calendar.title', "Calendar"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideMyEventsJson() async             => _provideResponseData(await Events2().loadEventsResponse(Events2Query(
-    types: { Event2TypeFilter.admin },
-    timeFilter: null,
-  )));
-
-  // rokwire.illinois.edu/gr
-
-  List<Widget> get _groupsContent => <Widget>[
+  Widget get _groupsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myGroups] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_groups.title', "My Groups"),
-      dataProvider: _provideMyGroupsJson,
+      key: _storedDataKeys[_StoredDataSource.groups] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.groups,
+      dataProvider: () => Groups().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.groups.title', "Groups"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _pollsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myGroupsStats] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_social_stats.title', "My Social Stats"),
-      dataProvider: _provideMySocialStatsJson,
+      key: _storedDataKeys[_StoredDataSource.polls] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.polls,
+      dataProvider: () => Polls().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.polls.title', "Polls"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideMyGroupsJson() async             => _provideResponseData(await Groups().loadUserGroupsResponse());
-  Future<String?> _provideMySocialStatsJson() async        => _provideResponseData(await Social().loadStatsResponse());
-
-  // rokwire.illinois.edu/polls
-
-  List<Widget> get _pollsContent => <Widget>[
+  Widget get _surveysContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myPools] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_polls.title', "My Polls"),
-      dataProvider: _provideMyPollsJson,
+      key: _storedDataKeys[_StoredDataSource.surveys] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataProvider: () => Surveys().loadUserDataJson(),
+      dataSource: _StoredDataSource.surveys,
+      title: Localization().getStringEx('panel.profile.stored_data.surveys.title', "Surveys"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _lmsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.participatedPolls] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.participated_polls.title', "Participated Polls"),
-      dataProvider: _provideParticipatedPollsJson,
+      key: _storedDataKeys[_StoredDataSource.lms] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataProvider: () => Canvas().loadUserDataJson(),
+      dataSource: _StoredDataSource.lms,
+      title: Localization().getStringEx('panel.profile.stored_data.lms.title', "Courses"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideMyPollsJson() async              => _provideResponseData(await Polls().getPollsResponse(PollFilter(
-    myPolls: true
-  )));
-  Future<String?> _provideParticipatedPollsJson() async    => _provideResponseData(await Polls().getPollsResponse(PollFilter(
-      respondedPolls: true
-  )));
-
-  // rokwire.illinois.edu/surveys
-
-  List<Widget> get _surveysContent => <Widget>[
+  Widget get _appointmentsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.mySurveys] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_surveys.title', "My Surveys"),
-      dataProvider: _provideMySurveysJson,
+      key: _storedDataKeys[_StoredDataSource.appointments] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.appointments,
+      dataProvider: () => Appointments().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.appointments.title', "Appointments"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _occupationsContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.participatedSurveys] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.participated_surveys.title', "Participated Surveys"),
-      dataProvider: _provideParticipatedSurveysJson,
+      key: _storedDataKeys[_StoredDataSource.occupations] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.occupations,
+      dataProvider: () => Occupations().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.occupations.title', "Occupations"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideMySurveysJson() async            => _provideResponseData(await Surveys().loadCreatorSurveysResponse());
-  Future<String?> _provideParticipatedSurveysJson() async  => _provideResponseData(await Surveys().loadUserSurveyResponsesResponse());
-
-  // rokwire.illinois.edu/identity
-
-  List<Widget> get _identityContent => <Widget>[
+  Widget get _transportationContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myStudentId] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_student_id.title', "My Student ID"),
-      dataProvider: _provideMyStudentIdJson,
+      key: _storedDataKeys[_StoredDataSource.transportation] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.transportation,
+      dataProvider: () => Transportation().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.transportation.title', "Transportation"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _wellnessContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myStudentClassification] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_student_classification.title', "My Student Classification"),
-      dataProvider: _provideMyStudentClassificationJson,
+      key: _storedDataKeys[_StoredDataSource.wellness] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.wellness,
+      dataProvider: () => Wellness().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.wellness.title', "Wellness"),
       updateController: _updateController,
-    ),
+    );
+
+  Widget get _assistantContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myMobileCredentials] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_mobile_credentials.title', "My Mobile Credentials"),
-      dataProvider: _provideMyMobileCredentialsJson,
+      key: _storedDataKeys[_StoredDataSource.assistant] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.assistant,
+      dataProvider: () => Assistant().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.assistant.title', "Assistant"),
       updateController: _updateController,
-    ),
-  ];
+    );
 
-  Future<String?> _provideMyStudentIdJson() async          => _provideResponseData(await Identity().loadStudentIdResponse());
-  Future<String?> _provideMyStudentClassificationJson() async => _provideResponseData(await Identity().loadStudentClassificationResponse());
-  Future<String?> _provideMyMobileCredentialsJson() async  => _provideResponseData(await Identity().loadMobileCredentialResponse());
-
-  // rokwire.illinois.edu/rewards
-
-  List<Widget> get _rewardsContent => <Widget>[
+  Widget get _gatewayContent =>
     _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myRewardsBalance] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_rewards_balance.title', "My Rewards Balance"),
-      dataProvider: _provideMyRewardsBalanceJson,
+      key: _storedDataKeys[_StoredDataSource.gateway] ??= GlobalKey<_ProfileStoredDataWidgetState>(),
+      dataSource: _StoredDataSource.gateway,
+      dataProvider: () => Gateway().loadUserDataJson(),
+      title: Localization().getStringEx('panel.profile.stored_data.gateway.title', "External Services"),
       updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myRewardsHistory] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_rewards_history.title', "My Rewards History"),
-      dataProvider: _provideMyRewardsHistoryJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideMyRewardsBalanceJson() async     => _provideResponseData(await Rewards().loadBalanceResponse());
-  Future<String?> _provideMyRewardsHistoryJson() async     => _provideResponseData(await Rewards().loadHistoryResponse());
-
-  // rokwire.illinois.edu/wellness
-
-  List<Widget> get _wellnessContent => <Widget>[
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myWellnessToDoCategories] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_wellness_todo_categories.title', "My Wellness ToDo Categories"),
-      dataProvider: _provideMyWellnessToDoCategoriesJson,
-      updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myWellnessToDoItems] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_wellness_todo_items.title', "My Wellness ToDo Items"),
-      dataProvider: _provideMyWellnessToDoItemsJson,
-      updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myWellnessRingsRecords] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_wellness_rings_records.title', "My Wellness Rigns Records"),
-      dataProvider: _provideMyWellnessRingsRecordsJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideMyWellnessToDoCategoriesJson() async => _provideResponseData(await Wellness().loadToDoCategoriesResponse());
-  Future<String?> _provideMyWellnessToDoItemsJson() async  => _provideResponseData(await Wellness().loadToDoItemsResponse());
-  Future<String?> _provideMyWellnessRingsRecordsJson() async => _provideResponseData(await WellnessRings().requestRingRecordsResponse());
-
-  // rokwire.illinois.edu/appointments
-
-  List<Widget> get _appointmentsContent => <Widget>[
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myAppointments] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_appointments.title', "My Appointments"),
-      dataProvider: _provideMyAppointmentsJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideMyAppointmentsJson() async       => _provideResponseData(await Appointments().loadAppointmentseResponse());
-
-  // rokwire.illinois.edu/occupations
-
-  List<Widget> get _occupationsContent => <Widget>[
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.myOccupationMatches] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.my_occupation_matches.title', "My Occupation Matches"),
-      dataProvider: _provideMyOccupationMatchesJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideMyOccupationMatchesJson() async  => _provideResponseData(await Occupations().getAllOccupationMatchesResponse());
-
-  // icard.uillinois.edu
-
-  List<Widget> get _icardContent => <Widget>[
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.iCard] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.i_card.title', "iCard"),
-      dataProvider: _provideICardJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideICardJson() async                => _provideResponseData(await Auth2().loadICardResponse());
-
-  // housing.illinois.edu
-
-  List<Widget> get _housingContent => <Widget>[
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.studentSummary] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.student_summary.title', "My Student Summary"),
-      dataProvider: _provideStudentSummaryJson,
-      updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.illiniCashBalance] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.illini_cash_balance.title', "My Illini Cash Balance"),
-      dataProvider: _provideIlliniCashBalanceJson,
-      updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.illiniCashTransactions] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.illini_cash_transactions.title', "My Illini Cash Transactions"),
-      hint: Localization().getStringEx('panel.profile.stored_data.label.this_year', " (this year)"),
-      dataProvider: _provideIlliniCashTransactionsJson,
-      updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.mealPlanTransactions] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.meal_plan_transactions.title', "My Meal Plan Transactions"),
-      hint: Localization().getStringEx('panel.profile.stored_data.label.this_year', " (this year)"),
-      dataProvider: _provideMealPlanTransactionsJson,
-      updateController: _updateController,
-    ),
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.cafeCreditTransactions] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.cafe_credit_transactions.title', "My Cafe Credit Transactions"),
-      hint: Localization().getStringEx('panel.profile.stored_data.label.this_year', " (this year)"),
-      dataProvider: _provideCafeCreditTransactionsJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideStudentSummaryJson() async       => _provideResponseData(await _provideStudentSummaryResponse());
-  Future<String?> _provideIlliniCashBalanceJson() async    => _provideResponseData(await IlliniCash().loadBalanceRequest());
-  Future<String?> _provideIlliniCashTransactionsJson() async => _provideResponseData(await IlliniCash().loadTransactionHistoryResponse(DateTime(DateTime.now().year), DateTime.now()));
-  Future<String?> _provideMealPlanTransactionsJson() async => _provideResponseData(await IlliniCash().loadMealPlanTransactionHistoryResponse(DateTime(DateTime.now().year), DateTime.now()));
-  Future<String?> _provideCafeCreditTransactionsJson() async => _provideResponseData(await IlliniCash().loadCafeCreditTransactionHistoryResponse(DateTime(DateTime.now().year), DateTime.now()));
-
-  Future<Response?> _provideStudentSummaryResponse() async {
-    dynamic result = await IlliniCash().loadStudentSummaryResponse();
-    return (result is Response) ? result : null;
-  }
-  
-  // recent_items
-
-  List<Widget> get recentItemsContent => <Widget>[
-    _ProfileStoredDataWidget(
-      key: _storedDataKeys[_StoredDataType.recentItems] ??= GlobalKey(),
-      title: Localization().getStringEx('panel.profile.stored_data.recent_items.title', "My Recently Viewed"),
-      dataProvider: _provideRecentItemsJson,
-      updateController: _updateController,
-    ),
-  ];
-
-  Future<String?> _provideRecentItemsJson() async => RecentItems.loadRecentItemsSource();
+    );
 
   // Header Bar
 
@@ -558,26 +319,21 @@ class _ProfileStoredDataPanelState extends State<ProfileStoredDataPanel> {
 
   // Implementation
 
-  String? _provideResponseData(Response? response) => ((response != null) && (response.statusCode >= 200) && (response.statusCode <= 301)) ?
-    (JsonUtils.encode(JsonUtils.decode(response.body), prettify: true) ?? response.body) : null;
-
   Future<void> _onRefresh() async {
-    _updateController.add(ProfileStoredDataPanel.notifyRefresh);
+    _updateController.add(ProfileStoredDataPanel._notifyRefresh);
   }
 
   void _onCopyAll() {
     Analytics().logSelect(target: 'Copy All');
 
     String combinedJson = "";
-    for (_StoredDataType dataType in _StoredDataType.values) {
-      GlobalKey? dataTypeKey = _storedDataKeys[dataType];
-      State<StatefulWidget>? state = dataTypeKey?.currentState;
-      _ProfileStoredDataWidgetState? dataTypeState = (state is _ProfileStoredDataWidgetState) ? state : null;
+    for (_StoredDataSource dataType in _StoredDataSource.values) {
+      _ProfileStoredDataWidgetState? dataTypeState = _storedDataKeys[dataType]?.currentState;
       if (dataTypeState != null) {
         if (dataTypeState.widget.title != null) {
-          combinedJson += '\n// ${dataTypeState.widget.title}\n';
+          combinedJson += '\n// ${dataTypeState._displayTitle}\n';
         }
-        combinedJson += dataTypeState._providedData ?? 'NA\n';
+        combinedJson += dataTypeState._displayContent ?? 'NA\n';
       }
     }
     Clipboard.setData(ClipboardData(text: combinedJson)).then((_) {
@@ -594,141 +350,282 @@ class _ProfileStoredDataPanelState extends State<ProfileStoredDataPanel> {
 
 
 class _ProfileStoredDataWidget extends StatefulWidget {
+  final _StoredDataSource dataSource;
+  final _UserDataProvider dataProvider;
   final String? title;
-  final String? hint;
-  final _StoredDataProvider dataProvider;
   final StreamController<String>? updateController;
   final EdgeInsetsGeometry margin;
 
   _ProfileStoredDataWidget({
     // ignore: unused_element
     super.key,
-    this.title, this.hint,
+    required this.dataSource,
     required this.dataProvider,
+    this.title,
     this.updateController,
     // ignore: unused_element
     this.margin = const EdgeInsets.symmetric(horizontal: 16, vertical: 12)
   });
+
+  String get titleKey => this.dataSource.toString();
 
   @override
   State<StatefulWidget> createState() => _ProfileStoredDataWidgetState();
 }
 
 class _ProfileStoredDataWidgetState extends State<_ProfileStoredDataWidget> {
-  TextEditingController _textController = TextEditingController();
-  bool _providingData = false;
-  String? _providedData;
+  bool _loading = false;
+  Map<String, dynamic>? _userData;
+  Map<String, String>? _displayUserData;
 
   @override
   void initState() {
 
     widget.updateController?.stream.listen((String command) {
-      if (command == ProfileStoredDataPanel.notifyRefresh) {
-        _refresh();
+      if (command == ProfileStoredDataPanel._notifyRefresh) {
+        _refreshUserData();
       }
     });
 
-    _init();
+    _initUserData();
+
     super.initState();
   }
 
   @override
   void dispose() {
-    _textController.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    Widget? headingWidget = _buildHeadingWidget();
     return Padding(padding: widget.margin, child:
-      Column(crossAxisAlignment: CrossAxisAlignment.start, mainAxisSize: MainAxisSize.min, children: [
-        if (headingWidget != null)
-          Padding(padding: EdgeInsets.only(bottom: 4), child: headingWidget,),
-        Stack(children: [
-          TextField(
-            maxLines: 5,
-            readOnly: true,
-            controller: _textController,
-            decoration: InputDecoration(
-              border: _textBorder,
-              focusedBorder: _textBorder,
-              contentPadding: EdgeInsets.symmetric(horizontal: 12, vertical: 8)
+      Column(crossAxisAlignment: CrossAxisAlignment.start, mainAxisSize: MainAxisSize.min, children:
+        _widgetContent
+      ),
+    );
+  }
+
+  List<Widget> get _widgetContent {
+    if (_loading) {
+      return _loadingContent;
+    }
+    else if (_userData == null) {
+      return _errorContent;
+    }
+    else if (_userData?.isEmpty == true) {
+      return _emptyContent;
+    }
+    else {
+      return _userDataContent;
+    }
+  }
+
+  List<Widget> get _loadingContent => <Widget>[
+    _ProfileStoredDataEntryWidget(
+      titleKey: widget.titleKey,
+      titleText: widget.title,
+      hintText: '...',
+      progress: true,
+    )
+  ];
+
+  List<Widget> get _errorContent => <Widget>[
+    _ProfileStoredDataEntryWidget(
+      titleKey: widget.titleKey,
+      titleText: widget.title,
+      //hintText: Localization().getStringEx("logic.general.error", "Error"),
+      contentText: Localization().getStringEx('widget.profile.stored_data.retrieve.failed.message', 'Failed to retrieve data.'),
+      error: true,
+    ),
+  ];
+
+  List<Widget> get _emptyContent => <Widget>[
+    _ProfileStoredDataEntryWidget(
+      titleKey: widget.titleKey,
+      titleText: widget.title,
+      contentText: Localization().getStringEx('widget.profile.stored_data.retrieve.empty.message', 'No stored information.'),
+    ),
+  ];
+
+  List<Widget> get _userDataContent {
+    List<Widget> entries = <Widget>[];
+    if (_userData != null) {
+      for (String entryKey in _userData!.keys) {
+
+        entries.add(Padding(padding: EdgeInsets.only(top: entries.isNotEmpty ? 24 : 0), child:
+          _ProfileStoredDataEntryWidget(
+            titleKey: widget.titleKey,
+            titleText: widget.title,
+            hintKey: entryKey,
+            contentText: _displayUserData?[entryKey],
+          ),
+        ));
+      }
+    }
+    return entries;
+  }
+
+
+  Future<void> _initUserData() async {
+    setState(() {
+      _loading = true;
+    });
+    Map<String, dynamic>? userData = await widget.dataProvider();
+    if (mounted) {
+      setState(() {
+        _displayUserData = _toUserData(_userData = userData);
+        _loading = false;
+      });
+    }
+  }
+
+  Future<void> _refreshUserData() async {
+    if (mounted) {
+      setState(() {
+        _loading = true;
+        _displayUserData = _toUserData(_userData = null);
+      });
+      Map<String, dynamic>? userData = await widget.dataProvider();
+      if (mounted) {
+        setState(() {
+          _displayUserData = _toUserData(_userData = userData);
+          _loading = false;
+        });
+      }
+    }
+  }
+
+  Map<String, String>? _toUserData(Map<String, dynamic>? sourceUserData) =>
+    sourceUserData?.map((key, value) => MapEntry(key, JsonUtils.encode(value, prettify: true) ?? ''));
+
+  String get _displayTitle =>
+    widget.title ?? Localization().getString('panel.profile.stored_data.source.${widget.titleKey}.title') ?? StringUtils.capitalize(widget.titleKey, allWords: true, splitDelimiter: '_', joinDelimiter: ' ');
+
+  String? get _displayContent => JsonUtils.encode(_userData);
+
+}
+
+class _ProfileStoredDataEntryWidget extends StatefulWidget {
+  final String? titleKey;
+  final String? titleText;
+
+  final String? hintKey;
+  final String? hintText;
+
+  final String? contentText;
+
+  final bool progress;
+  final bool error;
+
+  _ProfileStoredDataEntryWidget({
+    // ignore: unused_element
+    super.key,
+    this.titleKey, this.titleText,
+    this.hintKey, this.hintText,
+    this.contentText,
+    this.progress = false,
+    this.error = false,
+  });
+
+  @override
+  State<StatefulWidget> createState() => _ProfileStoredDataEntryWidgetState();
+}
+
+class _ProfileStoredDataEntryWidgetState extends State<_ProfileStoredDataEntryWidget> {
+  late TextEditingController _contentTextController;
+
+  @override
+  void initState() {
+    _contentTextController = TextEditingController(text: widget.contentText ?? '');
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    _contentTextController.dispose();
+    super.dispose();
+  }
+
+  @override
+  void didUpdateWidget(_ProfileStoredDataEntryWidget oldWidget) {
+    _contentTextController.text = widget.contentText ?? '';
+    super.didUpdateWidget(oldWidget);
+  }
+
+  @override
+  Widget build(BuildContext context) =>
+    Column(crossAxisAlignment: CrossAxisAlignment.start, mainAxisSize: MainAxisSize.min, children: [
+      _headingWidget,
+      Stack(children: [
+        _textContentWidget,
+        if (widget.progress)
+          Positioned.fill(child:
+            Align(alignment: Alignment.center, child:
+              _progressWidget,
             ),
-            style: (_providedData != null) ? Styles().textStyles.getTextStyle('widget.input_field.text.regular') : Styles().textStyles.getTextStyle('widget.item.small.thin.italic'),
           ),
 
-          Visibility(visible: _providingData, child:
-            Positioned.fill(child:
-              Align(alignment: Alignment.center, child:
-                SizedBox(height: 24, width: 24, child:
-                  CircularProgressIndicator(color: Styles().colors.fillColorSecondary, strokeWidth: 3),
-                ),
-              ),
+        if (!widget.progress && !widget.error && _contentTextController.text.isNotEmpty)
+          Positioned.fill(child:
+            Align(alignment: Alignment.topRight, child:
+              _copyButton,
             ),
           ),
+      ],),
+    ]);
 
-          Visibility(visible: !_providingData && StringUtils.isNotEmpty(_providedData), child:
-            Positioned.fill(child:
-              Align(alignment: Alignment.topRight, child:
-                InkWell(onTap: _onCopy, child:
-                  Padding(padding: EdgeInsets.all(12), child:
-                    Styles().images.getImage('copy', excludeFromSemantics: true),
-                  ),
-                ),
-              ),
-            ),
-          ),
-
-        ],),
+  Widget get _headingWidget {
+    String title = _displayTitle, hint = _displayHint;
+    return RichText(text:
+      TextSpan(style: Styles().textStyles.getTextStyle('widget.title.small.fat'), children: <InlineSpan>[
+        if (StringUtils.isNotEmpty(title))
+          TextSpan(text: StringUtils.isNotEmpty(hint) ? "${title} / "  : title, style: Styles().textStyles.getTextStyle('widget.title.small.fat')),
+        if (StringUtils.isNotEmpty(hint))
+          TextSpan(text: hint, style: Styles().textStyles.getTextStyle('widget.title.small.semi_fat.light')),
       ]),
     );
   }
 
-  Widget? _buildHeadingWidget() {
-    List<InlineSpan> spans = <InlineSpan>[
-      if (StringUtils.isNotEmpty(widget.title))
-        TextSpan(text: widget.title, style: Styles().textStyles.getTextStyle('widget.title.small.fat')),
-      if (StringUtils.isNotEmpty(widget.hint))
-        TextSpan(text: widget.hint, style: Styles().textStyles.getTextStyle('widget.title.small.semi_fat.light')),
-    ];
-    return spans.isNotEmpty ? RichText(
-      text: TextSpan(style: Styles().textStyles.getTextStyle('widget.title.small.fat'), children: spans)
-    ) : null;
-  }
+  String get _displayTitle =>
+    widget.titleText ?? Localization().getString('panel.profile.stored_data.source.${widget.titleKey}.title') ?? StringUtils.capitalize(widget.titleKey ?? '', allWords: true, splitDelimiter: '_', joinDelimiter: ' ');
 
-  InputBorder get _textBorder =>
+  String get _displayHint =>
+    widget.hintText ?? Localization().getString('panel.profile.stored_data.source..${widget.titleKey}.${widget.hintKey}.title') ?? StringUtils.capitalize(widget.hintKey ?? '', allWords: true, splitDelimiter: '_', joinDelimiter: ' ');
+
+  Widget get _textContentWidget => TextField(
+    maxLines: 5,
+    readOnly: true,
+    controller: _contentTextController,
+    decoration: _contentTextDecoration,
+    style: widget.error ? Styles().textStyles.getTextStyle('widget.input_field.text.regular') : Styles().textStyles.getTextStyle('widget.item.small.thin.italic'),
+  );
+
+  InputDecoration get _contentTextDecoration => InputDecoration(
+    border: _contentTextBorder,
+    focusedBorder: _contentTextBorder,
+    contentPadding: EdgeInsets.symmetric(horizontal: 12, vertical: 8)
+  );
+
+  InputBorder get _contentTextBorder =>
     OutlineInputBorder(borderSide: BorderSide(color: Colors.black, width: 1.0));
 
-  Future<void> _refresh() async {
-    if (mounted) {
-      setState(() {
-        _providedData = null;
-      });
-      _textController.text = '';
-      await _init();
-    }
-  }
+  Widget get _progressWidget =>
+    SizedBox(height: 24, width: 24, child:
+      CircularProgressIndicator(color: Styles().colors.fillColorSecondary, strokeWidth: 3),
+    );
 
-  Future<void> _init() async {
-    setState(() {
-      _providingData = true;
-    });
-    widget.dataProvider().then((String? data) {
-      if (mounted) {
-        setState(() {
-          _providedData = data;
-          _providingData = false;
-        });
-        _textController.text = data ?? Localization().getStringEx('widget.profile.stored_data.retrieve.failed.message', 'Failed to retrieve data');
-      }
-    });
-  }
+  Widget get _copyButton =>
+    InkWell(onTap: _onCopy, child:
+      Padding(padding: EdgeInsets.all(12), child:
+        Styles().images.getImage('copy', excludeFromSemantics: true),
+      ),
+    );
 
   void _onCopy() {
-    Analytics().logSelect(target: 'Copy All');
-    if (_providedData != null) {
-      Clipboard.setData(ClipboardData(text: _providedData ?? '')).then((_) {
+    Analytics().logSelect(target: 'Copy', source: "$_displayTitle / $_displayHint");
+    if (widget.contentText?.isNotEmpty == true) {
+      Clipboard.setData(ClipboardData(text:  widget.contentText ?? '')).then((_) {
         AppToast.showMessage(Localization().getStringEx('widget.profile.stored_data.copied.succeeded.message', 'Copied to your clipboard!'));
       });
     }